+++ src/ray_casting.c
@@ -6,7 +6,7 @@
 /*   By: waraissi <waraissi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/24 10:17:00 by waraissi          #+#    #+#             */
-/*   Updated: 2023/08/04 22:27:15 by waraissi         ###   ########.fr       */
+/*   Updated: 2023/08/05 18:11:32 by waraissi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ double	intersection_with_horizontal(t_win *vars, double i, double angle)
 	first_horizontal_intersection(vars, t, i);
 	while (true)
 	{
-		if (vars->r_cast->x_h < 0 || vars->r_cast->y_h < 0 || vars->r_cast->x_h / 50 > ft_strlen(vars->map[0]) || vars->r_cast->y_h / 50 > 13)
+		if (vars->r_cast->x_h < 0 || vars->r_cast->y_h < 0 || vars->r_cast->x_h / 50 > (int)ft_strlen(vars->map[0]) || vars->r_cast->y_h / 50 > 13)
 			break;
 		if (vars->map[(int)vars->r_cast->y_h / 50][(int)vars->r_cast->x_h / 50] == '1' || vars->map[(int)vars->r_cast->y_h / 50][(int)vars->r_cast->x_h / 50] == ' ' || !wall_coll_h(vars, angle))
 			break;
@@ -78,7 +78,7 @@ double	intersection_with_vertical(t_win *vars, double i, double angle)
 	first_vertical_intersection(vars, t, i);
 	while (true)
 	{
-		if (vars->r_cast->x_v < 0 || vars->r_cast->y_v < 0 || vars->r_cast->x_v / 50 > ft_strlen(vars->map[0]) || vars->r_cast->y_v / 50 > 13)
+		if (vars->r_cast->x_v < 0 || vars->r_cast->y_v < 0 || vars->r_cast->x_v / 50 > (int)ft_strlen(vars->map[0]) || vars->r_cast->y_v / 50 > 13)
 			break;
 		if (vars->map[(int)vars->r_cast->y_v / 50][(int)vars->r_cast->x_v / 50] == '1' || vars->map[(int)vars->r_cast->y_v / 50][(int)vars->r_cast->x_v / 50] == ' ' || !wall_coll_v(vars, angle))
 			break;
